Create an initial iteration of a CLI argument parser.
The parser accepts, upon construction,  a list of definitions which describe the kinds of
arguments to look for. It provides a parse function that parses the args
available through Lua's "arg" variable. It returns a table where keys
are argument names found when parsing, and values set to the values
provided on the command line.